requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
asyncio-throttle>=1.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
hypothesis>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0